selftests/bpf: Fix kmem_cache_iter test failure on 32-bit
The following test_progs errors are seen running on 32-bit armhf:

  root@qemu-armhf:/usr/libexec/kselftests-bpf# ./test_progs -a kmem_cache_iter/check_slabinfo
  test_kmem_cache_iter:PASS:kmem_cache_iter__open_and_load 0 nsec
  test_kmem_cache_iter:PASS:skel_attach 0 nsec
  test_kmem_cache_iter:PASS:iter_create 0 nsec
  9p-fcall-cache-1: 131096
  ext4_groupinfo_4k: 136
  [...]
  kmalloc-rcltest_kmem_cache_iter:PASS:read 0 nsec
  subtest_kmem_cache_iter_check_slabinfo:PASS:kmem_cache_lookup 0 nsec
  subtest_kmem_cache_iter_check_slabinfo:FAIL:kmem_cache_name unexpected kmem_cache_name: actual '9p-fcall-cache-1' != expected ''
  subtest_kmem_cache_iter_check_slabinfo:PASS:kmem_cache_objsize 0 nsec
  subtest_kmem_cache_iter_check_slabinfo:PASS:kmem_cache_lookup 0 nsec
  subtest_kmem_cache_iter_check_slabinfo:FAIL:kmem_cache_name unexpected kmem_cache_name: actual 'ext4_groupinfo_4k' != expected ''
  [...]
  subtest_kmem_cache_iter_check_slabinfo:PASS:kmem_cache_seen_eq 0 nsec
  #148/2   kmem_cache_iter/check_slabinfo:FAIL
  #148     kmem_cache_iter:FAIL

The failure results from a stack buffer overflow due to an inconsistent
struct layout between the 32-bit userspace host and 64-bit BPF VM.

Both prog_tests/kmem_cache_iter.c and progs/kmem_cache_iter.c define type

  struct kmem_cache_result {
         char name[SLAB_NAME_MAX];
         long obj_size;
  };

for accessing map 'slab_result', but the struct is smaller in userspace
due to the use of 'long', and overflows the receive buffer during a call
to bpf_map_lookup_elem(). The above errors are from overwriting the string
'name' in this particular case.

Fix by using type 'unsigned int' for kmem_cache_result:obj_size, matching
also the source field from 'struct kmem_cache'.

Fixes: a496d0cdc84d ("selftests/bpf: Add a test for kmem_cache_iter")

+++ tools/testing/selftests/bpf/prog_tests/kmem_cache_iter.c
@@ -10,7 +10,7 @@
 
 struct kmem_cache_result {
 	char name[SLAB_NAME_MAX];
-	long obj_size;
+	unsigned int obj_size;
 };
 
 static void subtest_kmem_cache_iter_check_task_struct(struct kmem_cache_iter *skel)
@@ -32,7 +32,7 @@ static void subtest_kmem_cache_iter_check_slabinfo(struct kmem_cache_iter *skel)
 	FILE *fp;
 	int map_fd;
 	char name[SLAB_NAME_MAX];
-	unsigned long objsize;
+	unsigned int objsize;
 	char rest_of_line[1000];
 	struct kmem_cache_result r;
 	int seen = 0;
@@ -50,7 +50,7 @@ static void subtest_kmem_cache_iter_check_slabinfo(struct kmem_cache_iter *skel)
 	fscanf(fp, "# %*s %*s %*s %*s %*s %*s : %[^\n]\n", rest_of_line);
 
 	/* Compare name and objsize only - others can be changes frequently */
-	while (fscanf(fp, "%s %*u %*u %lu %*u %*u : %[^\n]\n",
+	while (fscanf(fp, "%s %*u %*u %u %*u %*u : %[^\n]\n",
 		      name, &objsize, rest_of_line) == 3) {
 		int ret = bpf_map_lookup_elem(map_fd, &seen, &r);
